Added tiles for hedges and tried to change the background to a tilemap.
It didn't work to to copy and paste the tilemaps to the background game object, they couldn't be used as tilemaps. So I kept them in the grid gameobject but they didn't work with the background script so I made a bigger white background to use for the background object that were used before.